msm_vidc: Fix hang due to incorrect buffer unmap
Buffers queued with the same index share extrdata plane
info in dynamic buffer mode as well. When a buffer
is dequeued while another buffer(queued earlier with the
same index)is held in driver due to fw ref, dqbuf ends
up unmapping the old buffer, as the buffer_info used
to unregister is the one used to access the extradata
uvaddr, which happens to map to the old buffer as it is
earlier in the list, and the extradata plane is processed
last.

Fix the issue by populating the plane uvaddr in reverse
order, so that the correct buffer_info is unmapped.

Change-Id: I15723cc13c582fd5970641d51ba46456935ab4fb
Signed-off-by: Surajit Podder <spodder@codeaurora.org>