hrtimer: don't send an IPI unnecessarily during migration
commit 1c68c8be8195 ("timer/hrtimer: reprogram remote CPU's
timer hardware") added the required logic to kick the target
CPU during migration. The value returned by the enqueue_hrtimer()
is used to decide if the target CPU timer hardware reprogramming
is required or not. But It only tells if it the left most timer
in the queued base not the cpu_base->expires_next.tv64. So, we
end up sending an IPI unnecessarily.

The hrtimer_check_target() check inside switch_hrtimer_base()
does not allow migration in the first place, if the current
timer expires before the programmed next event on the target
CPU. So there is no need to send an IPI.

Change-Id: I80341acfee89d607b21d71bdc45f043b6bdc8609
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>