net: sch_generic: Fix deadlock involving dev_watchdog()
There is a deadlock in dev_watchdog() if there is no watchdog timeout
set, which is often the case.

Fix this by hold the lock first then check for watchdog timeout, note
that we put it at the end of the check to get qdisc_tx_is_noop()
get a completed cycle.

Signed-off-by: Wang Han <416810799@qq.com>